SI-4694 Adapt function interfaces to 64 bit addressing

+++ include/vmem/vmem_ram.h
@@ -10,8 +10,8 @@
 
 #include <vmem/vmem.h>
 
-void vmem_ram_read(vmem_t * vmem, uint32_t addr, void * dataout, uint32_t len);
-void vmem_ram_write(vmem_t * vmem, uint32_t addr, const void * datain, uint32_t len);
+void vmem_ram_read(vmem_t * vmem, uint64_t addr, void * dataout, intptr_t len);
+void vmem_ram_write(vmem_t * vmem, uint64_t addr, const void * datain, intptr_t len);
 
 typedef struct {
 	void * physaddr;
